# contracts/openapi/parslex-v0.yaml
openapi: 3.1.0
info:
  title: ParsLex Platform API
  version: 1.1.0
  description: |
    Enterprise Domain-Adaptive Legal AI Platform — REST API v0.

    Phase 1 covers identity, documents, collections, assistant sessions, and audit.
    All endpoints require authentication unless noted.
  contact:
    name: ParsLex Architecture Team
servers:
  - url: http://localhost:8000/api/v1
    description: Local development
  - url: https://parslex.local/api/v1
    description: On-premise deployment
...
paths:
...
  /organizations/{orgId}/documents:
    get:
      tags: [Documents]
      summary: List documents in organization
      operationId: listDocuments
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
        - name: document_type
          in: query
          schema:
            $ref: '#/components/schemas/DocumentType'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Paginated document list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentListResponse'
    post:
      tags: [Documents]
      summary: Upload a new document
      operationId: uploadDocument
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title]
              properties:
                file:
                  type: string
                  format: binary
                title:
                  type: string
                document_type:
                  $ref: '#/components/schemas/DocumentType'
                language:
                  type: string
                  enum: [fa, en, bilingual]
                workspace_id:
                  type: string
                  format: uuid
                metadata:
                  type: string
                  description: JSON-encoded metadata object
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          $ref: '#/components/responses/BadRequest'
  /organizations/{orgId}/documents/{documentId}:
    get:
      tags: [Documents]
      summary: Get document by ID
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Documents]
      summary: Soft-delete document
      operationId: deleteDocument
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '204':
          description: Document deleted
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/documents/{documentId}/versions:
    get:
      tags: [Documents]
      summary: List document versions
      operationId: listDocumentVersions
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Version list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/DocumentVersion'
  /organizations/{orgId}/documents/{documentId}/download:
    get:
      tags: [Documents]
      summary: Download latest document version
      operationId: downloadDocument
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'
  /organizations/{orgId}/collections:
    get:
      tags: [Collections]
      summary: List knowledge collections
      operationId: listCollections
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: workspace_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Collection list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/KnowledgeCollection'
    post:
      tags: [Collections]
      summary: Create knowledge collection
      operationId: createCollection
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCollectionRequest'
      responses:
        '201':
          description: Collection created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeCollection'
  /organizations/{orgId}/collections/{collectionId}:
    get:
      tags: [Collections]
      summary: Get collection by ID
      operationId: getCollection
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/CollectionId'
      responses:
        '200':
          description: Collection details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeCollection'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Collections]
      summary: Update collection
      operationId: updateCollection
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCollectionRequest'
      responses:
        '200':
          description: Collection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeCollection'
  /organizations/{orgId}/collections/{collectionId}/documents:
    post:
      tags: [Collections]
      summary: Add document to collection
      operationId: addDocumentToCollection
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/CollectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [document_id]
              properties:
                document_id:
                  type: string
                  format: uuid
      responses:
        '204':
          description: Document added
    delete:
      tags: [Collections]
      summary: Remove document from collection
      operationId: removeDocumentFromCollection
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/CollectionId'
        - name: document_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Document removed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    CollectionId:
      name: collectionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
...
    DocumentType:
      type: string
      enum:
        - contract
        - regulation
        - law
        - template
        - policy
        - court_decision
        - book
        - tender
        - other
    DocumentStatus:
      type: string
      enum:
        - pending
        - processing
        - ready
        - failed
    Document:
      type: object
      required: [id, organization_id, title, document_type, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
        document_type:
          $ref: '#/components/schemas/DocumentType'
        language:
          type: string
          enum: [fa, en, bilingual]
        status:
          $ref: '#/components/schemas/DocumentStatus'
        current_version:
          $ref: '#/components/schemas/DocumentVersion'
        metadata:
          type: object
          additionalProperties: true
        created_by:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DocumentVersion:
      type: object
      required: [id, document_id, version_number, file_name, file_size, content_hash]
      properties:
        id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        version_number:
          type: integer
        file_name:
          type: string
        file_size:
          type: integer
        content_type:
          type: string
        content_hash:
          type: string
        storage_key:
          type: string
        created_at:
          type: string
          format: date-time
    DocumentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Document'
        total:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
    KnowledgeCollection:
      type: object
      required: [id, organization_id, name, created_at]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        legal_domain_id:
          type: string
        document_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateCollectionRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        workspace_id:
          type: string
          format: uuid
        legal_domain_id:
          type: string
    UpdateCollectionRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
...